refactoring: replace dirs
