+++ Sources.cmake
@@ -1,0 +1,5 @@
+set(ALGORITHMS_ON_STRINGS_LIB
+  algorithms_on_strings/longest_common_substring/LongestSubstring.cpp
+  algorithms_on_strings/longest_common_substring/LongestSubstring.h
+)
+
